+++ src/Stack/Types/Config.hs
@@ -101,6 +101,8 @@ data Config =
          -- ^ --extra-lib-dirs arguments
          ,configConfigMonoid        :: !ConfigMonoid
          -- ^ @ConfigMonoid@ used to generate this
+         ,configConcurrentTests     :: !Bool
+         -- ^ Run test suites concurrently
          }
 
 -- | Information on a single package index
@@ -432,6 +434,8 @@ data ConfigMonoid =
     -- ^ See: 'configExtraIncludeDirs'
     ,configMonoidExtraLibDirs        :: !(Set Text)
     -- ^ See: 'configExtraLibDirs'
+    ,configMonoidConcurrentTests     :: !(Maybe Bool)
+    -- ^ See: 'configConcurrentTests'
     }
   deriving Show
 
@@ -452,6 +456,7 @@ instance Monoid ConfigMonoid where
     , configMonoidJobs = Nothing
     , configMonoidExtraIncludeDirs = Set.empty
     , configMonoidExtraLibDirs = Set.empty
+    , configMonoidConcurrentTests = Nothing
     }
   mappend l r = ConfigMonoid
     { configMonoidDockerOpts = configMonoidDockerOpts l <> configMonoidDockerOpts r
@@ -470,6 +475,7 @@ instance Monoid ConfigMonoid where
     , configMonoidJobs = configMonoidJobs l <|> configMonoidJobs r
     , configMonoidExtraIncludeDirs = Set.union (configMonoidExtraIncludeDirs l) (configMonoidExtraIncludeDirs r)
     , configMonoidExtraLibDirs = Set.union (configMonoidExtraLibDirs l) (configMonoidExtraLibDirs r)
+    , configMonoidConcurrentTests = configMonoidConcurrentTests l <|> configMonoidConcurrentTests r
     }
 
 instance FromJSON ConfigMonoid where
@@ -493,6 +499,7 @@ instance FromJSON ConfigMonoid where
          configMonoidJobs <- obj .:? "jobs"
          configMonoidExtraIncludeDirs <- obj .:? "extra-include-dirs" .!= Set.empty
          configMonoidExtraLibDirs <- obj .:? "extra-lib-dirs" .!= Set.empty
+         configMonoidConcurrentTests <- obj .:? "concurrent-tests"
          return ConfigMonoid {..}
 
 -- | Newtype for non-orphan FromJSON instance.