+++ ChangeLog.md
@@ -14,6 +14,7 @@
 * Rename `stack docker exec` to `stack exec --plain`
 * Add the `--skip-msys` flag [#377](https://github.com/commercialhaskell/stack/issues/377)
 * `--keep-going`, turned on by default for tests and benchmarks [#478](https://github.com/commercialhaskell/stack/issues/478)
+* `concurrent-tests: BOOL` [#492](https://github.com/commercialhaskell/stack/issues/492)
 
 ## 0.1.1.0